Fix dynamic file name
